Generate list of tags and store in a json file
A function has been added to update app which generates a list of
tags in the project from the metadata.json file. A new property
is defined in the settings file which stores the path to the tags
json file. A tags.json file has already been generated for the
existing metadata.json.

Resolves #39.

Related issues: #21, #40.