Use python decouple to separate settings from code
With python decouple, all the settings are kept in a .env file
and an example .env.example has been provided as a reference.
Necessary change has been provided in the doucmentation and also
a dependency has been added in the requirements.txt. Also a command
for copying .env.example as .env has been added to the .travis.yml
so that build does not fail.

Resolves #36