update master (#50)
* add IDEA folder to gitignore (#18)

* fix titles, dynamic searchresults title depending on query

* update titles

* Develop (#32)

* static files setting

* sample images

* template for random selection

* added image

* debug false

* serving images through github and removed staticfiles app

* code of conduct

* update readme

* fix doc order error on install section

* Suggestions based on Search query.

* Made required changes

* Develop (#31)

* add IDEA folder to gitignore (#18)

* fix titles, dynamic searchresults title depending on query

* update titles

* Add virtualenv command to Readme file

* Generate list of tags and store in a json file

A function has been added to update app which generates a list of
tags in the project from the metadata.json file. A new property
is defined in the settings file which stores the path to the tags
json file. A tags.json file has already been generated for the
existing metadata.json.

Resolves #39.

Related issues: #21, #40.

* Update the metadata.json

* Update the tags.json file

The tags.json file is updated according to the new metadata.json
file and `src` and `test` are excluded from coming into the tags.
Also '-' is replaced with ' ' in the update_tags function.