Add flake8 code linter
Flake8 configuration file has been created and the existing
violations have been solved. Flake8 is also added to the
requirements and a script has been added to the travis.yml file
so that code is checked for linting issues after it is pushed.

**The violations that has been added to ignore section of the
flake8 configuration file has to be solved in near future. This is
just a temporary solution.**

Resolves #49